requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.1
importlib-metadata==0.23
lxml==4.4.1
more-itertools==7.2.0
namedlist==1.7
numpy==1.17.4
packaging==19.2
pandas==0.25.3
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.5
pytest==5.2.2
python-dateutil==2.8.1
pytz==2019.3
selenium==3.141.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
xlrd==1.2.0
zipp==0.6.0
